requirements.txt
appdirs==1.4.3
argcomplete==1.8.2
base58==0.2.4
boto3==1.4.4
botocore==1.5.40
click==6.7
Django==1.8
django-filter==1.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
django-webpack-loader==0.5.0
django-forms-bootstrap==3.1.0
docutils==0.13.1
durationpy==0.4
future==0.16.0
futures==3.1.1
hjson==2.0.2
jmespath==0.9.2
kappa==0.6.0
lambda-packages==0.14.2
Markdown==2.6.8
packaging==16.8
placebo==0.8.1
PyJWT==1.5.0
PyMySQL==0.7.11
pyparsing==2.2.0
python-dateutil==2.6.0
slackclient==1.0.7
python-slugify==1.2.4
PyYAML==3.12
requests==2.14.2
s3transfer==0.1.10
six==1.10.0
toml==0.9.2
tqdm==4.11.2
troposphere==1.9.3
Unidecode==0.4.20
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.41.1
